Removed extraneous __mocks__ dir, moved unit test files to folders that match the project file structure
